Day and Night Demolisher
This modlet allows Demolishers to spawn in the Wasteland during the daytime and at nighttime. The Demolishers have simply been added to the requisite zombie group for that location and they will spawn within standard probability parameters. However, in real play testing, being placed at the top of the group list seems to give the Demolisher a higher probability of spawning first or very early into the spawn cycle.